Refactor code
* Put `RacingcarModel` into `domain` package.
* Changed `calculateMaxPosition` method impl to use collection API.
* Added missing package declaration in `StringCalculator` tests.
