+++ src/main/java/com/woowacourse/javaracingcar/RacingcarController.java
@@ -1,5 +1,7 @@
 package com.woowacourse.javaracingcar;
 
+import com.woowacourse.javaracingcar.domain.RacingcarModel;
+import com.woowacourse.javaracingcar.domain.RandomNumberGenerator;
 import com.woowacourse.javaracingcar.interfaces.NumberGenerator;
 import com.woowacourse.javaracingcar.interfaces.UserInterface;
