+++ src/test/java/com/woowacourse/stringcalculator/StringCalculatorTest.java
@@ -1,3 +1,5 @@
+package com.woowacourse.stringcalculator;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
