+++ src/test/java/com/woowacourse/stringcalculator/StringCalculatorUtilTest.java
@@ -1,3 +1,5 @@
+package com.woowacourse.stringcalculator;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -5,6 +7,7 @@ import java.util.LinkedList;
 import java.util.Queue;
 
 import static org.assertj.core.api.Assertions.assertThat;
+
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class StringCalculatorUtilTest {
